# Fix sampling in sort-merge join