This PR addresses #1208. Previously, the sampling (apart from being plain wrong) was using the frequency of values to determine splits in the data. Since the chance of sampling the same value multiple times is low (and even then, it does not need to be a good split value), the splitting resulted in highly skewed clusters.

Now, we gather a small `SampleRequest` in which each materialized segment puts some samples. Then, the sample requests are merged, sorted and we select the required number of split values.

I measured the performance with an active scheduler and the change is minimal. I expected it to be better since the values are now (almost) equally sized.

Now:
```
➜  OpossumDB git:(fix/sort_merge_join) ✗ ./rel/hyriseMicroBenchmarks  --benchmark_filter="BM_Sam" --benchmark_repetitions=10
Generating TPC-H data set with scale factor 1 and Dictionary encoding:
[...]
----------------------------------------------------------------------------------------
Benchmark                                                 Time           CPU Iterations
----------------------------------------------------------------------------------------
[...]
TPCHDataMicroBenchmarkFixture/BM_Sampling_mean   1468207936 ns 1006052700 ns          1
TPCHDataMicroBenchmarkFixture/BM_Sampling_median 1464570165 ns 1000790500 ns          1
TPCHDataMicroBenchmarkFixture/BM_Sampling_stddev   23104783 ns   11344929 ns          1
```

Before
```
➜  OpossumDB_master git:(master) ✗ ./rel/hyriseMicroBenchmarks  --benchmark_filter="BM_Sam" --benchmark_repetitions=10
Generating TPC-H data set with scale factor 1 and Dictionary encoding:
[...]
----------------------------------------------------------------------------------------
Benchmark                                                 Time           CPU Iterations
----------------------------------------------------------------------------------------
[...]
TPCHDataMicroBenchmarkFixture/BM_Sampling_mean   1502313768 ns 1017044400 ns          1
TPCHDataMicroBenchmarkFixture/BM_Sampling_median 1515102442 ns 1017976500 ns          1
TPCHDataMicroBenchmarkFixture/BM_Sampling_stddev   68923687 ns   15745084 ns          1
```